next: typed::Stream
