+++ ohkami/src/response/_test.rs
@@ -138,8 +138,8 @@ async fn test_stream_response() {
 
     let res = Response::OK()
         .with_stream(
-            repeat_by(3, |i| Result::<Vec<u8>, std::convert::Infallible>::Ok(
-                format!("This is message#{i} !").into_bytes()
+            repeat_by(3, |i| Result::<_, std::convert::Infallible>::Ok(
+                format!("This is message#{i} !")
             ))
         )
         .with_headers(|h| h
