+++ ohkami/src/typed/mod.rs
@@ -3,4 +3,7 @@ pub mod status;
 mod payload;
 pub use payload::{Payload, PayloadType};
 
+#[cfg(feature="sse")]
+mod stream;
+
 pub use ohkami_macros::{Payload, Query};
